[Binary Translation] New version of ICDFG (more standardized)
This new version (contained in instruction/cdfg/wip) follows the more
general way of generating CDFGs, were instead of a graph that contains
all of the nodes, the CDFG is composed of DFGs and control nodes. The
new implementation also tries to be more general, so if another CDFG is
needed, it should be (mostly) easy to implement it. Also the JGrapht
DOTexporter was rewritten because it didn't support any subgraphs, and
it didn't allow being extended (a lot of methods and subclasses were
private).

Known issues:
	- The generation of nested ifs produces wrong control flow
	- Still no implementation of subscripts and functions
	- Code is still very messy in some parts
	- The CDFG still doesn't have exceptions for when the generator
produces a wrong result (needed for generalization)
	- Still needs to be more generalized