[Binary Translation] Added Riscv full flow and tried to improve SSAPass
NOTE: The SSA pass is still not functional, so segments and complex
 instructions are not still supported ! The whole package needs to be
rewritten and maybe a more generic SSA pass needs to be applied. Also
the RISCV custom instructions currently use one opcode for each of the
parts of the usage of the generated modules, but maybe this should be
changed (RISCV 32bit only has something like 30 free opcodes, and the
base extensions already occupy half of those opcodes).